@W-19368426: [MSDK] Rewrite Android Screen Lock (a.k.a Passcode) for RTL UI (Peer-Review Cleanup)
